requirements.txt
discord.py
colorama
pymongo
